Break out prep for clipping and save output locally
